refactor: move ToastType enum to enums file
