+++ src/components/Toast/enums.ts
@@ -1,0 +1,6 @@
+export enum ToastType {
+  Regular = 'regular',
+  Success = 'success',
+  Error = 'error',
+  Loading = 'loading',
+}
